Reorgonized streamlit code to new files.
